requirements.txt
python-telegram-bot
urllib3
bs4
lxml